[misc] add explicit serialVersionUID, mark non-serializable fields as transient, expand wildcard imports, and add missing @SuppressWarnings annotations

+++ src/main/java/org/mariadb/jdbc/util/ClientParser.java
@@ -389,14 +389,14 @@ public final class ClientParser implements PrepareResult {
         case 'V':
           if (state == LexState.Normal
               && valuesBracketPositions.isEmpty()
-              && (lastChar == ')' || ((byte) lastChar <= 40))
+              && (lastChar == ')' || (lastChar <= 40))
               && queryLength > i + 7
               && equalsIgnoreCase(query[i + 1], (byte) 'a')
               && equalsIgnoreCase(query[i + 2], (byte) 'l')
               && equalsIgnoreCase(query[i + 3], (byte) 'u')
               && equalsIgnoreCase(query[i + 4], (byte) 'e')
               && equalsIgnoreCase(query[i + 5], (byte) 's')
-              && (query[i + 6] == '(' || ((byte) query[i + 6] <= 40))) {
+              && (query[i + 6] == '(' || (query[i + 6] <= 40))) {
             afterValues = true;
             if (query[i + 6] == '(') {
               valuesBracketPositions.add(i + 6);
